Convert line numbers to NSUInteger / integerValue
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/projectcenter/trunk@38021 72102866-910b-0410-8b05-ffd578937521

+++ Framework/PCProjectBuilder.m
@@ -1324,9 +1324,10 @@
 				   windowed:NO];
   if (editor)
     {
+      // TODO / FIXME using a NSPoint here is weak since it is Float vs. integer line numbers
       position = NSPointFromString([error objectForKey:@"Position"]);
       [projectEditor orderFrontEditorForFile:[error objectForKey:@"File"]];
-      [editor scrollToLineNumber:(unsigned int)position.x];
+      [editor scrollToLineNumber:(NSUInteger)position.x];
 
 /*      NSLog(@"%i: %@(%@): %@", 
 	    position.x,
